Introduce WASM client using web-sys
This change introduces the counterpart to the native client that can
work in WASM land. In a browser environment, we can access Web APIs
(and the Fetch API in particular) to issue HTTP requests. Perhaps most
importantly, this change jumps through all the hoops required to get
this functionality tested properly.