Add support for providing a non-empty body
So far we have strictly worked with empty request bodies. While most
likely sufficient for commonly used GET requests, the moment we want to
POST something we are likely out of luck.
To fix this shortcoming, this change adds support for accepting and
passing through a proper body. Note that right now we only support
a string to be provided and not an arbitrary byte sequence. We have not
yet found a way to feed arbitrary bytes to the request issued through
the Fetch API in a WASM environment.