Add a way to directly create a Client object from the underlying client
Our client object is just a wrapper around the underlying client. As
such, it makes sense to be able to create it from such an object. To
that end, this change introduces from conversions that allow for that to
happen. Similarly, users may desire to eventually get the underlying
client object back. Hence, we also provide the inverse functionality for
extracting it back out.
Note that ideally we would not have a struct type at all but would just
provide a trait from this crate that is implemented for the supported
client types. Unfortunately, that vision is complicated by the fact that
we are working with async functions, which cannot trivially be used in
traits (but would require support by async-trait, for example).