Add logic to install wasm-bindgen-test-runner locally
In some circumstances it is not possible or desired to have a
system-installed wasm-bindgen-test-runner. In a CI environment, for
example, we are bound to having to install it manually, because the
binary is typically not packaged by distributions.
To that end, this change adds the logic to install it locally in a
temporary location. Furthermore, we adjust the test target to use such a
locally installed binary if one is available.