Introduce Issue trait
This change introduces a trait encapsulating the ability to issue an
HTTP request to the crate: Issue. In the future, this trait will
supersede the current Client type.
There are two main advantages to using a trait and just implementing it
for the respective clients we aim to work with:
1) it means that no back-and-forth conversions are necessary
2) we are much more flexible with respect to generic arguments on the
   underlying clients, such as the connector that is used on hyper
   clients and which enables HTTPS usage