Make native client Issue::issue futures implement Send
It has been a long standing problem that for <reasons> pretty much any
WASM future does not implement Send. As such, there is no way for us to
make the Issue trait's async issue method return a future that
implements Send. However, because the trait is also implemented for the
native client, we have inherited this property to its implementation.
That in turn, is a potential cause of problems, as it means that code
using the crate in a purely native context is now penalized. Hence,
there is an inherent tension between implementing Send for the native
client and providing different guarantees for different compilation
targets or providing the same guarantees (as we have done so far) but
making native clients potentially pay a price.
With this change we switch away from the uniform model and make the
Issue trait return futures that implement Send when compiling natively.
The reasoning for this step is two-fold:
1) issues caused by different trait guarantees can be identified at
  compile time (as opposed to runtime testing, which is much harder to
  do for the different supported environments), and we encourage users
  to build for the targets of interest.
2) the hope is that eventually WASM futures will implement Send and we
  can make Issue trait generated futures Send unconditionally.