Pin wasm-bindgen dependency to 0.2.74
It has become borderline impossible to maintain a stable and predictable
CI build, because with every bump of wasm-bindgen we hit new snags
building it. To provide an end to the quest of finding stable versions
that we can use and for which we can build the corresponding
test-runner, this change pins the version we consume to 0.2.74.