Refactor test setup
This change reworks our test setup slightly. Specifically, we adjust the
test.sh script to only invoke one test for a given target, which can be
provided as a positional parameter (and rename it to test-target.sh).
Furthermore, we split the existing test target into test-native and
test-wasm, each running the corresponding test suite.
This split will allow us to run the tests separately in CI, which will
allow us to both parallelize and cache more and ultimately speed up
runs.