Do not set RUSTFLAGS for WASM jobs
From the looks of it, wasm-bindgen does not honor the RUSTFLAGS
environment variable correctly. As a result of that, unnecessary
rebuilds are triggered for a supposed change of this variable and that
happens over and over again.
To work around this issue, this change unsets the RUSTFLAGS variable in
all WASM jobs.