Make native client TLS aware
Most web services, sites, etc., use encrypted HTTPS instead of
unencrypted HTTP these days. We need to be able to support encrypted
connections to cater to these use cases. The WASM client using Web APIs
already supports that, but the native client needs a special TLS
connector to get there.
With this change we adjust the native client to use such a TLS connector
in order to get support for working with HTTPS enabled services. Note
that this really only applies to the already deprecated Client struct.
Users of the Issue trait have full control over which connector to use
already. However, we still add a test for this functionality, which
provides additional value.