Forcefully install wasm-bindgen-test-runner binary
Our wasm-bindgen-test-runner installation step is failing occasionally
with Cargo complaining that the binary is already installed. The way it
was supposed to work was that cargo should skip installation, because a
binary is already present.
Given that we already only perform compilation conditionally, there is
no harm in just forcefully installing the executable and so with this
change that is what we do.