Eliminate explicit test-runner installation step
Eliminate the need for manually installing the wasm-bindgen-test-runner
on the file system in order to run WASM tests. The way we get around
this requirement is by creating an executable that just wraps the
functionality via a crates.io import and then using 'cargo run --example
test-runner' as the runner for the wasm32-unknown-unknown target.